refactor: change the folder structure
