+++ src/app/Components/NavBar/index.tsx
@@ -1,14 +1,13 @@
 "use client";
 import React, { useState } from "react";
 import { usePathname } from "next/navigation";
-import NavItem from "./NavItem";
-import MenuIcon from "../assets/MenuIcon";
-import CloseIcon from "../assets/CloseIcon";
+import NavItem from "../NavItem";
+import MenuIcon from "../../assets/MenuIcon";
+import CloseIcon from "../../assets/CloseIcon";
 
 const MenuItems = [
   {
     path: "/", label: "Home"},
-
    {path: "/bookings", label: "Bookings"} ,	 
   { path: "/about", label: "About Us"},
 ]
