+++ src/app/Components/NavItem/index.tsx
@@ -1,13 +1,7 @@
 "use client";
 import React from "react";
 import Link from "next/link";
-
-interface NavItemProps {
-  href: string;
-  pathname: string;
-  children: React.ReactNode;
-  onClick?: () => void;
-}
+import { NavItemProps } from "@/app/types/NavItems.types";
 
 const NavItem: React.FC<NavItemProps> = ({ href, pathname, children }) => {
   const isActive = pathname === href;
@@ -17,7 +11,7 @@ const NavItem: React.FC<NavItemProps> = ({ href, pathname, children }) => {
       href={href}
       className={`hover:underline ${isActive ? "underline text-blue-600" : ""}`}
     >
-      {" "}
+     
       {children}
     </Link>
   );
