+++ src/app/types/NavItems.types.ts
@@ -1,0 +1,6 @@
+export type NavItemProps = {
+    href: string;
+    pathname: string;
+    children: React.ReactNode;
+    onClick?: () => void;
+  };
